requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
decouple==0.0.7
distro==1.9.0
Django==5.1.6
django-allauth==65.4.1
django-allauth-ui==1.6.1
django-recaptcha==4.0.0
django-rosetta==0.10.1
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
jwt==1.3.1
nltk==3.9.1
normalize==2.1.0
numpy==2.2.3
openai==1.63.2
pandas==2.2.3
pillow==11.1.0
polib==1.2.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
richenum==2.0.2
rosetta==0.3
scipy==1.15.2
six==1.17.0
slippers==0.6.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
